# server/openapi.yaml
openapi: 3.0.3
info:
  title: Sylex Assistant Pro API
  description: |
    Cloud backend API for Sylex Assistant Pro - providing ASR (Automatic Speech Recognition),
    NLU/Chat, and TTS (Text-to-Speech) services for in-vehicle voice assistant.
    
    This API is designed for automotive integration with focus on:
    - Low latency (<300ms TTFB, <1.5s E2E)
    - Safety-first design
    - Multilingual support (DE primary)
    - Streaming support for real-time interaction
  version: 1.0.0
  contact:
    name: Sylex Assistant Pro Team
    email: support@car-assistant-pro.com
  license:
    name: Proprietary
servers:
  - url: https://api.car-assistant-pro.com/v1
    description: Production server
  - url: https://staging-api.car-assistant-pro.com/v1
    description: Staging server
  - url: http://localhost:3000/v1
    description: Local development server
...
paths:
...
  /asr:
    post:
      tags:
        - ASR
      summary: Convert speech to text
      description: |
        Transcribes audio input to text using cloud ASR provider.
        Supports multiple audio formats and languages.
        Returns transcript with confidence score.
      operationId: speechToText
      requestBody:
        required: true
        content:
          multipart/form-data:
            schema:
              type: object
              properties:
                audio:
                  type: string
                  format: binary
                  description: Audio file (WAV, MP3, FLAC, OGG)
                language:
                  type: string
                  description: BCP-47 language code
                  default: de-DE
                  example: de-DE
                  enum:
                    - de-DE
                    - en-US
                    - en-GB
                    - fr-FR
                enableProfanityFilter:
                  type: boolean
                  default: true
                enableAutomaticPunctuation:
                  type: boolean
                  default: true
                model:
                  type: string
                  description: ASR model variant
                  default: default
                  enum:
                    - default
                    - automotive
                    - command
              required:
                - audio
      responses:
        '200':
          description: Transcription successful
          headers:
            X-Request-ID:
              schema:
                type: string
              description: Unique request identifier for tracing
            X-Processing-Time-Ms:
              schema:
                type: integer
              description: Server processing time in milliseconds
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/AsrResponse'
        '400':
          description: Bad request (invalid audio format, missing required fields)
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '413':
          description: Payload too large (audio file exceeds size limit)
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '429':
          description: Rate limit exceeded
          headers:
            X-RateLimit-Limit:
              schema:
                type: integer
              description: Request limit per time window
            X-RateLimit-Remaining:
              schema:
                type: integer
              description: Remaining requests in current window
            X-RateLimit-Reset:
              schema:
                type: integer
              description: Unix timestamp when rate limit resets
            Retry-After:
              schema:
                type: integer
              description: Seconds to wait before retry
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '500':
          description: Internal server error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '503':
          description: Service unavailable (ASR provider down)
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
  /asr/stream:
    post:
      tags:
        - ASR
      summary: Stream speech to text (WebSocket upgrade)
      description: |
        Establishes WebSocket connection for streaming ASR.
        Client sends audio chunks, server returns partial and final transcripts.
      operationId: speechToTextStream
      parameters:
        - name: Upgrade
          in: header
          required: true
          schema:
            type: string
            enum: [websocket]
        - name: Connection
          in: header
          required: true
          schema:
            type: string
            enum: [Upgrade]
      responses:
        '101':
          description: Switching Protocols to WebSocket
        '400':
          description: Bad request
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
...
components:
  schemas:
...
    AsrResponse:
      type: object
      properties:
        requestId:
          type: string
          format: uuid
          description: Unique request identifier
        transcript:
          type: string
          description: Recognized text from audio
          example: "Navigate to nearest charging station"
        confidence:
          type: number
          format: float
          description: Confidence score (0.0 to 1.0)
          minimum: 0.0
          maximum: 1.0
          example: 0.95
        language:
          type: string
          description: Detected language
          example: "de-DE"
        alternatives:
          type: array
          description: Alternative transcriptions
          items:
            type: object
            properties:
              transcript:
                type: string
              confidence:
                type: number
                format: float
        processingTimeMs:
          type: integer
          description: Processing time in milliseconds
          example: 287
      required:
        - requestId
        - transcript
        - confidence
        - language
        - processingTimeMs
...
    ErrorResponse:
      type: object
      properties:
        error:
          type: object
          properties:
            code:
              type: string
              description: Error code
              enum:
                - INVALID_REQUEST
                - INVALID_AUDIO_FORMAT
                - AUDIO_TOO_LARGE
                - TEXT_TOO_LONG
                - RATE_LIMIT_EXCEEDED
                - UNAUTHORIZED
                - PROVIDER_ERROR
                - INTERNAL_ERROR
                - SERVICE_UNAVAILABLE
              example: "INVALID_AUDIO_FORMAT"
            message:
              type: string
              description: Human-readable error message
              example: "Unsupported audio format. Supported formats: WAV, MP3, FLAC, OGG"
            details:
              type: object
              description: Additional error details
              additionalProperties: true
            requestId:
              type: string
              format: uuid
              description: Request ID for tracking
          required:
            - code
            - message
      required:
        - error
  securitySchemes:
    BearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
      description: JWT token for authentication
security:
  - BearerAuth: []
...